refactor: change structure of the renderCards
