+++ src/components/render/createCard.ts
@@ -1,8 +1,7 @@
 import { ConstantsDom } from '../../models/Dom';
 import { DataProducts } from '../../interfaces/Data';
 import Card from '../elements/CreateCard';
-import { createElement, elementDomStorage } from '../elements/generateElement';
-import { LocalStorage } from '../../utils/persistentStorage';
+import { createElement } from '../elements/generateElement';
 
 export function createCard(data: DataProducts) {
   const cardShell: Card = new Card(ConstantsDom.DIV, data, {
@@ -74,18 +73,4 @@ export function createCard(data: DataProducts) {
   });
   document.body.append(cardShell.elem);
   return cardShell;
-}
-
-export function renderCards(data: DataProducts[]): HTMLElement[] {
-  elementDomStorage.get(ConstantsDom.CARD_CONTAINER);
-  const arrayData: Card[] = [];
-  const arrayDataCard: HTMLElement[] = [];
-  data.forEach((cardData) => {
-    const card = createCard(cardData);
-    arrayData.push(card);
-    arrayDataCard.push(card.elem);
-  });
-  const CardsListStorage = new LocalStorage();
-  CardsListStorage.setItem('cards', arrayData);
-  return arrayDataCard;
 }
